+++ Wrappers/Python/cil/optimisation/utilities/callbacks.py
@@ -145,7 +145,7 @@ class EarlyStoppingObjectiveValue(Callback):
 
     Note
     -----
-    This callback only compares the last two calculated objective values. f update_objective_interval is greater than 1, the objective value is not calculated at each iteration (which is the default behaviour), only every update_objective_interval iterations.
+    This callback only compares the last two calculated objective values. If `update_objective_interval` is greater than 1, the objective value is not calculated at each iteration (which is the default behaviour), only every `update_objective_interval` iterations.
     
         '''
     def __init__(self, threshold=1e-6):
@@ -159,16 +159,16 @@ class EarlyStoppingObjectiveValue(Callback):
                 
 class CGLSEarlyStopping(Callback):
     '''Callback to work with CGLS. It causes the algorithm to terminate if  :math:`||A^T(Ax-b)||_2 < \epsilon||A^T(Ax_0-b)||_2` where `epsilon` is set to default as '1e-6', :math:`x` is the current iterate and :math:`x_0` is the initial value. 
-    It will also terminate if the algorithm begins to diverge i.e. if :math:`||x||_2> \Omega`, where `omega` is set to default as 1e6. 
+    It will also terminate if the algorithm begins to diverge i.e. if :math:`||x||_2> \omega`, where `omega` is set to default as 1e6. 
     Parameters
     ----------
     epsilon: float, default 1e-6 
         Usually a small number: the algorithm to terminate if :math:`||A^T(Ax-b)||_2 < \epsilon||A^T(Ax_0-b)||_2`
-    lambda: float, default 1e6 
-        Usually a large number: the algorithm will terminate if  :math:`||x||_2> \Omega`
+    omega: float, default 1e6 
+        Usually a large number: the algorithm will terminate if  :math:`||x||_2> \omega`
         
     Note
-    ---------
+    -----
     This callback is implemented to replicate the automatic behaviour of CGLS in CIL versions <=24. It also replicates the behaviour of https://web.stanford.edu/group/SOL/software/cgls/. 
     '''
     def __init__(self, epsilon=1e-6, omega=1e6):